Create Blackjack Highest.cpp
Output below (sum of card values is less than 21), above (sum of card values exceed 21), or blackjack (sum of card values add up to 21) and the highest card in your hand in relation to whether or not you have blackjack.  If there is an ace in your hand, but the sum will exceed 21, the ace must be counted as a 1.  The order of importance in face cards is jack, queen, then king.